bugfixes and magicsquare save and load

+++ Lab3/Lab3/game_base.cpp
@@ -44,7 +44,7 @@ void gameBase::getGame(int argc, char* argv[]) {
 	}
 	/*Neither Case*/
 	else {
-		throw badGameName;
+		throw (int)badGameName;
 	}
 }
